torrent: Don't attempt to rm -rf (null) on unload
  $ nbdkit torrent --dump-plugin
  [...]
  sh: -c: line 1: syntax error near unexpected token `('
  sh: -c: line 1: `rm -rf (null)'

This happened because the plugin attempts to delete a local cache
directory in .unload.  However when using --dump-plugin (or in other
cases) the directory had not been created and the cache variable was
NULL.

Fixes: commit 3cdd6e27e596a27fdb2f3d75286d484e59636939
(cherry picked from commit 7a7d3b491e7e058b4b3bfaccd89df4102d97284a)

+++ plugins/torrent/torrent.cpp
@@ -134,7 +134,7 @@ torrent_unload (void)
    * TMPDIR then it could break, but that's controlled by the person
    * running nbdkit.
    */
-  if (clean_cache_on_exit) {
+  if (clean_cache_on_exit && cache) {
     CLEANUP_FREE char *cmd;
     if (asprintf (&cmd, "rm -rf %s", cache) >= 0) {
 #pragma GCC diagnostic push
